This is a piece of coursework, please judge it accordingly. Updates may be slow or non-existent as a result.
Reducing the hits and misses and/or false alarms and correct rejections causes an error that I currently don't know how to fix. I consulted the TquanT version (link in code) and that also has an error under similar circumstances (with a swanky custom error message).

I struggled to get working user-defined functions to work, hence the repeated code and giant for loop. I know for loops can be stressful on the old computer, and can possibly go wrong and get you stuck infinitely. That's why iterations are locked at an upper limit of 10k and an empty data frame is defined prior to the initiation of the loop. Hopefully this minimises the risk.

None of my testing has run up against these problems thus far- feel free to test it to destruction and email problems/fixes to me or fix it yourself.

It is also a bit ugly.